refactor: move search_attribute logic to temporal_script.py
